Particle Stuff WIP
Ingen större skillnad.

+++ MortalPortal/Particles.h
@@ -48,7 +48,7 @@ public:
 
 	ID3D11UnorderedAccessView* getUAV();
 	unsigned int GetNrOfParticles();
-	int particleCounter;
+	float particleCounter;
 	std::vector<Particles> particle;
 	ConstantBufferData constantBufferData;
